backend/requirements.txt
a2a-sdk==0.3.22
ag-ui-protocol==0.1.10
agent-framework==1.0.0b260130
agent-framework-a2a==1.0.0b260130
agent-framework-ag-ui==1.0.0b260130
agent-framework-anthropic==1.0.0b260130
agent-framework-azure==0.0.0
agent-framework-azure-ai==1.0.0b260130
agent-framework-azure-ai-search==1.0.0b260130
agent-framework-azurefunctions==1.0.0b260130
agent-framework-chatkit==1.0.0b260130
agent-framework-copilotstudio==1.0.0b260130
agent-framework-core==1.0.0b260130
agent-framework-declarative==1.0.0b260130
agent-framework-devui==1.0.0b260130
agent-framework-durabletask==1.0.0b260130
agent-framework-github-copilot==1.0.0b260130
agent-framework-lab==1.0.0b251024
agent-framework-mem0==1.0.0b260130
agent-framework-ollama==1.0.0b260130
agent-framework-purview==1.0.0b260130
agent-framework-redis==1.0.0b260130
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.79.0
anyio==4.12.1
asyncio==4.0.0
attrs==25.4.0
azure-ai-agents==1.2.0b5
azure-ai-projects==2.0.0b3
azure-common==1.1.28
azure-core==1.38.0
azure-functions==1.25.0b3
azure-functions-durable==1.5.0
azure-identity==1.26.0b1
azure-search-documents==11.7.0b2
azure-storage-blob==12.29.0b1
backoff==2.2.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
cryptography==46.0.5
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
durabletask==1.3.0
durabletask.azuremanaged==1.3.0
fastapi==0.128.7
frozenlist==1.8.0
furl==2.1.4
github-copilot-sdk==0.1.24rc0
google-api-core==2.29.0
google-auth==2.49.0.dev0
googleapis-common-protos==1.72.0
griffe==1.15.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.13.0
jsonpath-ng==1.7.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mcp==1.26.0
mem0ai==1.0.3
microsoft-agents-activity==0.8.0.dev2
microsoft-agents-copilotstudio-client==0.8.0.dev2
microsoft-agents-hosting-core==0.8.0.dev2
ml_dtypes==0.5.4
msal==1.35.0b1
msal-extensions==1.3.1
multidict==6.7.1
numpy==2.4.2
ollama==0.6.1
openai==2.17.0
openai-agents==0.8.3
openai-chatkit==1.6.0
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
orderedmultidict==1.0.2
packaging==26.0
pluggy==1.6.0
ply==3.11
portalocker==3.2.0
posthog==7.8.5
powerfx==0.0.34
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.5.2
pydantic_core==2.41.5
PyJWT==2.11.0
pytest==8.3.3
pytest-cov==7.0.0
pytest-timeout==2.4.0
pytest-xdist==3.8.0
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-ulid==3.1.0
pythonnet==3.0.5
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
redis==7.1.1
redisvl==0.14.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.1.0b1
sse-starlette==3.2.0
starlette==0.52.1
tenacity==9.1.4
tqdm==4.67.3
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.5
yarl==1.22.0
zipp==3.23.0
